+++ HISTORY.md
@@ -2,6 +2,11 @@
 
 ---
 
+## 10.2.4
+
+`FIXED` tree render emptyicon
+`FIXED` force check right fixed if columns is changed
+
 ## 10.2.3
 
 * `FIXED` a logic addition to method `checkFixedMaxWidth` if offset is larger than 0